+++ packages/context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/context
 
+## 0.29.6
+
+### Patch Changes
+
+- @tangleai/models@0.29.6
+
 ## 0.29.5
 
 ### Patch Changes