+++ packages/gmpl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/gmpl
 
+## 0.29.6
+
+### Patch Changes
+
+- @tangleai/mas@0.29.6
+
 ## 0.29.5
 
 ### Patch Changes