+++ packages/jaren/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/jaren
 
+## 0.29.6
+
+### Patch Changes
+
+- @tangleai/agents@0.29.6
+  - @tangleai/context@0.29.6
+  - @tangleai/models@0.29.6
+
 ## 0.29.5
 
 ### Patch Changes